Replaced generation of the W vector high mode random fields with a generic API that also provides an interface to manipulate the 4D fermion sources and solutions pre and post inversion.
The original approach of having independent random numbers for the two flavor diagonal elements of the source is now an implementation of this.

Added an implementation of the above that has the random numbers for the two flavor diagonal elements related by complex conjugation. The implementation rotates the 4D sources in spin and flavor to X-conjugate form suitable for inversion with the X-conjugate Dirac operator. This rotation is now undone post-hoc. This implementation is 2x faster than the original.

Added a wrapper implementation of the 5D inverter API that performs intermediate conversion to and from X-conjugate form, to allow inversion with the X-conjugate Dirac operator

The setWhRandom method of the W vector has been removed. The random fields are now set by external functionality.

In computeVWhigh, the W random fields are now only set if this has not previously been performed

Added an EvecInterface implementation for when no evecs have been generated

Added a function to test that a 2f field is X-conjugate

Added implementations of operator- and operator*(scalar,matrix) for CPSsquareMatrix

Added a number of verification tests for the new X-conjugate rotation/inverse method