Moved the CPSfield alloc policies to a separate header in utils so they can be generally employed
To each alloc policy, one can now specify a preferred initial allocation location (host/device) which is used where it makes sense to control the setup. CPSfield now uses this option to default-allocate on the host rather than baking it into the allocator.

Added a "vector with a view" class that uses the alloc policies and offers view semantics to a memory region + unit tests

In vMv and vv mesonfield linalg operations, replaced the multi-GB temp buffers with the vector-with-a-view such that the memory manager can prevent these from causing the device to run out of memory